requirements.txt
accelerate==1.11.0
acvl_utils==0.2.5
annotated-types==0.7.0
appdirs==1.4.4
asttokens==3.0.1
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blosc2==3.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
connected-components-3d==3.26.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==4.4.2
diffusers==0.35.2
dotmap==1.3.30
dynamic_network_architectures==0.4.2
einops==0.8.1
executing==2.2.1
fft-conv-pytorch==1.2.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
gensim==4.4.0
gitdb==4.0.12
gitPython==3.1.45
graphviz==0.21
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
imagecodecs==2025.11.11
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpickle==4.1.1
jupyter_client==8.6.3
jupyter_core==5.9.1
#karateclub==1.3.0
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.3
linecache2==1.0.0
lpips==0.1.4
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==2.6.3
nibabel==5.3.2
nltk==3.9.2
nnunetv2==2.6.2
numexpr==2.14.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyGSP==0.6.1
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
python-louvain==0.16
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
pyzmq==27.1.0
# Editable install with no version control (radgraph==0.1.18) 
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-image
scikit-learn
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.44.0
simpleitk==2.5.2
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.22
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
torchxrayvision==1.4.0
tornado==6.5.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unittest2==1.1.0
urllib3==2.5.0
wandb==0.22.3
wcwidth==0.2.14
wrapt==2.0.1
yacs==0.1.8
zipp==3.23.0
